requirements.txt
pandas
numpy
scikit-learn
python-dotenv
requests
websockets
joblib
kerykeion
ecdsa
eth-account
